backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-dotenv==1.0.0
pillow==10.1.0
reportlab==4.0.7
cryptography==41.0.7
pyzipper==0.3.6
openpyxl==3.1.2
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# MySQL Support
pymysql==1.1.0
